fix: Remove misplaced curly brace, update README
This repo was a huge help to get me sorted out, and having ssh access to
my windows desktop without a login is like a dream come true. I found
the task failing due to a misplaced curly brace, and thought it would help
to add some details to the readme after setting up on a freshly installed
Windows 10 system.

Thanks for making this available.